added pch file to speed up compilation, refactored, established expose namespace
