+++ api/boostpython/api_geo_cell_data.cpp
@@ -1,11 +1,5 @@
+#include "boostpython_pch.h"
 
-
-
-#include <boost/python/def.hpp>
-#include <boost/python/args.hpp>
-#include <boost/python/class.hpp>
-#include <boost/python/copy_const_reference.hpp>
-#include <boost/python/return_value_policy.hpp>
 #include "core/geo_point.h"
 #include "core/geo_cell_data.h"
