+++ api/boostpython/api_precipitation_correction.cpp
@@ -1,6 +1,4 @@
-#include <boost/python/def.hpp>
-#include <boost/python/args.hpp>
-#include <boost/python/class.hpp>
+#include "boostpython_pch.h"
 
 #include "core/precipitation_correction.h"
