+++ api/boostpython/api_priestley_taylor.cpp
@@ -1,8 +1,4 @@
-
-
-#include <boost/python/def.hpp>
-#include <boost/python/args.hpp>
-#include <boost/python/class.hpp>
+#include "boostpython_pch.h"
 
 #include "core/priestley_taylor.h"
