+++ api/boostpython/boostpython_pch.h
@@ -1,0 +1,24 @@
+#if defined(_WINDOWS)
+#pragma once
+#endif
+
+#include <boost/python/module.hpp>
+#include <boost/python/def.hpp>
+#include <boost/python/args.hpp>
+#include <boost/python/class.hpp>
+#include <boost/python/scope.hpp>
+#include <boost/python/suite/indexing/vector_indexing_suite.hpp>
+#include <boost/python/suite/indexing/map_indexing_suite.hpp>
+#include <boost/python/overloads.hpp>
+#include <boost/python/return_internal_reference.hpp>
+#include <boost/python/copy_const_reference.hpp>
+#include <boost/python/handle.hpp>
+#include <boost/python/numeric.hpp>
+#include <boost/python/tuple.hpp>
+#include <boost/python/enum.hpp>
+#include <boost/python/operators.hpp>
+#include <boost/python/overloads.hpp>
+#include <boost/operators.hpp>
+
+
+#include "core/core_pch.h"
